Finalized done, missing check cycle of going online after ir

+++ services/spotsOnlineService/internal/handlers/httphdl/request.go
@@ -2,5 +2,5 @@ package httphdl
 
 type SpotDateRequest struct {
 	DurationApproximated int64 `json:"durationApproximated" binding:"required"`
-	MaximunCapacity      int   `json:"maximunCapacity" binding:"required"`
+	MaximunCapacity      int64 `json:"maximunCapacity" binding:"required"`
 }
